% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2b.R
\name{a2b}
\alias{a2b}
\title{Convert data format}
\usage{
a2b(a)
}
\arguments{
\item{a}{An object of class data.frame with the format a (one column for each size interval).}
}
\value{
b An object of class data.frame wih the format b (One column for sizes and one for abundance).
}
\description{
Convert fish transect data from format a to format b
}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>#'
}

