% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2c.R
\name{a2c}
\alias{a2c}
\title{Convert data format}
\usage{
a2c(a)
}
\arguments{
\item{a}{An object of class data.frame with the format a (one column for each size interval).}
}
\value{
c An object of class data.frame with the format c (one column for size and one row for each occurrence).
}
\description{
Convert fish transect data from format a to c
}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>
}

