% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2a.R
\name{b2a}
\alias{b2a}
\title{Convert data format}
\usage{
b2a(b)
}
\arguments{
\item{a}{A data.frame with the format b. Un data.frame en formato b}
}
\value{
A data.frame wih the format a. Un data.frame en formato a
}
\description{
Convert fish transect data from format b to format a; See details. Convierte datos de transectos de peces del formato b al formato a; Ver detalles
}
\examples{
#library(MPAtools)

#load a "b" sample dataset
data("b")

#convert a to b format
a <- b2a(b)
a

}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
