% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_model.R
\name{bio_model}
\alias{bio_model}
\title{Fit a generalized linear model}
\usage{
bio_model(data, covars = c("Temperatura", "Visibilidad", "Profundidad"))
}
\arguments{
\item{data}{A data.frame containing at least the columns for Indicador, Ano, Zona, Post, Temperature, Visibility, and Profundidad. The indicator can be any of the biophysical indicators generated by shannon(), density(), fish_biomass(), trophic(), fish_size()}

\item{covars}{a character vector specifying which covariates to keep. Options are Temperatura, Profundidad, Visibilidad or None (to drop all covariates).}
}
\value{
An object of class lm containing the fitted model
}
\description{
This fits a linear model with the form of Indicador ~ Ano + Zona + Post * Zona + Temperatura + Visibilidad + Profundidad to the data. Depending on the specified covars, Temperature, Visibility and Depth can be dropped from the model. Use define_covars to identify which covariates to keep or specify them manually.
}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
