% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_results.R
\name{bio_results}
\alias{bio_results}
\title{Results}
\usage{
bio_results(values, data, res, con)
}
\arguments{
\item{data}{The data to be analyzed}

\item{res}{The name of the reserve}

\item{con}{The name of the control site}

\item{input}{The input list from a Shiny application}
}
\value{
results, a tibble containing the name of the indicator as well as DiD estimates, p values, a string to be used as input for valueBox, the color of the valueBox, and the lm object fit to the indicator extracted from the data
}
\description{
Analyzes all the required indicators at once
}

