% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_results_i.R
\name{bio_results_i}
\alias{bio_results_i}
\title{Biophysical Results}
\usage{
bio_results_i(values, data, res, con)
}
\arguments{
\item{values}{A list, created within the Shiny app, that contains the list of indicators selected, the community, any objective species, and the year of implementation of the reserve}

\item{data}{A data.frame containing the raw data to be analyzed}

\item{res}{A string specifying the name of the reserve site(s)}

\item{con}{A string specifying the name of the control site(s)}
}
\value{
results, a tibble containing the name of the indicator as well as DiD estimates, p values, a string to be used as input for valueBox, the color of the valueBox, and the lm object fit to the indicator extracted from the data
}
\description{
Analyzes all the required biophysical indicators (for invertebrates) at once
}
\author{
Villasenor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
