% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_score.R
\name{bio_score}
\alias{bio_score}
\title{Scoring system}
\usage{
bio_score(model)
}
\arguments{
\item{model}{a linear model generated by the turfeffect function, where type has been set to "bio"}
}
\value{
score
}
\description{
Scoring system
}
