% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2a.R
\name{c2a}
\alias{c2a}
\title{Convert data format}
\usage{
c2a(c)
}
\arguments{
\item{a}{A data.frame with the format c. Un data.frame en formato c}
}
\value{
A data.frame wih the format a. Un data.frame en formato a
}
\description{
Convert fish transect data from format c to format a; See details. Convierte datos de transectos de peces del formato c al formato a; Ver detalles
}
\examples{
#library(MPAtools)

#load an "c" sample dataset
data("c")

#convert c to a format
a <- c2a(c)
a

}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
