% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2b.R
\name{c2b}
\alias{c2b}
\title{Convert data format}
\usage{
c2b(c)
}
\arguments{
\item{c}{An object of class data.frame wih the format c (One column for sizes and one for abundance).}
}
\value{
b An object of class data.frame with the format b (one column for each size interval).
}
\description{

}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>
}

