% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_wrapper.R
\name{covars_wrapper}
\alias{covars_wrapper}
\title{A wrapper for define_covars}
\usage{
covars_wrapper(data, year.imp)
}
\arguments{
\item{data}{dataset}

\item{year.imp}{year of implementation}
}
\value{
a vector stating which covariates to be used
}
\description{
Identifies for each sampling point (before-after-control-reserve) which covariates to use using define_covars. It then identifies the common covariates available for each sampling point and returns a string that bio_model can use to implement the correct model
}
