% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATAisFORMAT.R
\name{DATAisFORMAT}
\alias{DATAisFORMAT}
\title{DATA is FORMAT}
\usage{
DATAisFORMAT(data, format)
}
\arguments{
\item{data}{a data.frame that will be compared to a data.frame with the specified format as specified by the COBIApp}

\item{format}{The specified format to which data will be compared}
}
\value{
message an error message displayed if the data is not consistent with format
}
\description{
Evaluates if a passed data.frame has a specified format
}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>
}
