% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did}
\alias{did}
\title{Perform Difference-in-Difference analysis}
\usage{
did(data, year.imp)
}
\arguments{
\item{data}{A dataframe that has been created with any of the following functions: richness(), density(), fish_biomass(), fish_size(), trophic(). All these functions are part of this package.}

\item{year.imp}{An integer that indicates the year when the reserve was first implemented.}
}
\value{
Returns a linear model, where the value of the interactive term between Zone and Year is the difference in difference value.
}
\description{
Performs a difference-in-difference analysis on a series of data and compares the effect of the reserve.
}

