% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es50.R
\name{es50_timeblock}
\alias{es50_timeblock}
\title{Calculate es50 metric for an mpa for specific time points}
\usage{
es50_timeblock(mpa, key, age = 10, step_size = 10, verbose = interactive())
}
\arguments{
\item{mpa}{a row from wdpa table containing one mpa with metadata, obs and geometry column
If the obs column is not present then it is fetched from OBIS}

\item{key}{the wdpaid of the mpa}

\item{age}{numeric, ages to compute es50 for}

\item{step_size}{numeric duration in each step of analysis}

\item{verbose}{logical, if TRUE output helpful messages}
}
\value{
single row table containing same columns as mpa but with added timeblock columns
}
\description{
Calculate es50 metric for an mpa for specific time points
}
