% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es50.R
\name{es_intervals}
\alias{es_intervals}
\title{Calculate es50 metric for an mpa for specific time windows}
\usage{
es_intervals(
  mpa,
  key,
  ages = c(10, 20),
  step_size = 10,
  sample_size = 50,
  verbose = interactive()
)
}
\arguments{
\item{mpa}{a single row from wdpa table containing one mpa with metadata, obs and geometry column
If the obs column is not present then it is fetched from OBIS}

\item{key}{the wdpaid of the mpa (ignored)}

\item{ages}{numeric, one or more 'ages' relative to 'STATUS_YR' to compute es50 for}

\item{step_size}{numeric duration (in years) for each age (interval = age + time_step)}

\item{sample_size}{numeric see \code{\link[vegan]{rarefy}} \code{sample} argument.  50 computes ES50.}

\item{verbose}{logical, if TRUE output helpful messages}
}
\value{
tibble of ES statistics
}
\description{
Calculate es50 metric for an mpa for specific time windows
}
