% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{fetch_mpa}
\alias{fetch_mpa}
\title{Fetch MPA data by country from WDPA.  Be patient with this - it can take a while.}
\usage{
fetch_mpa(
  name = "Cuba",
  path = rappdirs::user_data_dir("wdpar"),
  ext = ".gpkg",
  clean = TRUE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{name}{character, either accepted country name or ISO-3 code. You can also
downlaod the global dataset using \code{global}.}

\item{path}{character, the output path.  We default to that specified by
\code{\link[rappdirs]{user_data_dir}}}

\item{ext}{character, the file extension, by default we use ".gpkg"}

\item{clean}{logical, if TRUE run the resut through \code{\link[wdpar]{wdpa_clean}}}

\item{overwrite}{logical, if TRUE overwrite exisiting datasets}

\item{...}{other arguments for , see \code{\link[wdpar]{wdpa_fetch}}}
}
\value{
sf table of protected areas
}
\description{
Fetch MPA data by country from WDPA.  Be patient with this - it can take a while.
}
\examples{
\dontrun{
x <- fetch_mpa(name = 'Cuba')
plot((x \%>\% dplyr::filter(MARINE == "1"))['REP_AREA'])
}
}
