% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{fetch_obis_country}
\alias{fetch_obis_country}
\title{Fetch OBIS data by country}
\usage{
fetch_obis_country(name = "Cuba", overwrite = TRUE, ...)
}
\arguments{
\item{name}{character, the name to save the MPA under (no spaces please!)}

\item{overwrite}{logical, if TRUE allow existing files to be overwritten}

\item{...}{other arguments for obis_fetch_mpa}
}
\value{
tibble of OBIS observations, possibly empty
}
\description{
No error checking is done to ensure the country folderand WDPA file
}
