% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_biomass.R
\name{fish_biomass}
\alias{fish_biomass}
\title{Calculate biomass}
\usage{
fish_biomass(data, ab = NULL, location, species = NULL)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Ano, Zonificacion (e.g. "Zona de pesca" or "Zona de no pesca"), Transecto, GeneroEspecie.}

\item{ab}{A dataframe that contains the a and b allometric parameters for each species. Dataframe must have three columns: GeneroEspecie, a, b}

\item{location}{A quoted string that indicates the location.}

\item{species}{A quoted string that indicates a species for which density should be calulated.}
}
\value{
size A dataframe with columns for Ano, Zonificacion, Transect Number, Species, and fish biomass.
}
\description{
Calulates biomass for a selected species or the entire dataset, by year and transect, for each location.
}

