% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_size.R
\name{fish_size}
\alias{fish_size}
\title{Calculate a descriptor for fish size}
\usage{
fish_size(data, site, stat = "mean", species = NULL)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Year, Zone (e.g. reserve or control site), Transect Number, GenusSpecies.}

\item{site}{A quoted string that indicates the site. Options are "Rosario", "IslaNatividad", "IslaMagdalena".}

\item{stat}{A quoted string that indicates which statistic should be calculated. Options include "max", "min", "mean", and "median". User can also use "all" to calculate all of the above.}

\item{species}{A quoted string that indicates a species for which density should be calulated. Common cases include "Paralabrax clathratus", "Semicossipjus pulcher", or "Anisotremus davidsonii".}
}
\value{
size A dataframe containing year, zone (insido or outside the reserve), transect number, species, and the statistic required.
}
\description{
Calulates mean, median, maximum, or minimum size (cm) for a list of species (or selected species, as indicated by parameter species) by year, and transect, for a speciic location.
}

