% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_na.R
\name{fix_na}
\alias{fix_na}
\title{Fix NA}
\usage{
fix_na(data)
}
\arguments{
\item{data}{a data.frame, generated by any of the MPAtools::indicator functions}
}
\value{
a data.frame where all-na variables (i.e. the entire column had NA) are replaced with a 1
}
\description{
Removes all na-variables (see value) to avoid errors when performing the linear model
}

