% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_score.R
\name{gen_score}
\alias{gen_score}
\title{General Score}
\usage{
gen_score(percent)
}
\arguments{
\item{percent}{A number, between 0 and 100 that represents the percentage of positive indicators}
}
\value{
a color
}
\description{
According to the percentage of indicators that are positive, return a color based on a red - orange - yellow - green - olive scale, define by 20% intervals
}

