% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{has_mpa}
\alias{has_mpa}
\title{Determine if the local storage contains a set of MPAs by name(s)}
\usage{
has_mpa(name = c("Cuba", "Mongolia"), ...)
}
\arguments{
\item{name}{character, the name(s) of the country or 'global' that you have previously
downloaded}

\item{...}{arguments for \code{list_mpa}}
}
\value{
one or more logicals as a named vector
}
\description{
Determine if the local storage contains a set of MPAs by name(s)
}
