% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indB_sel.R
\name{indB_sel}
\alias{indB_sel}
\title{Select indicators}
\usage{
indB_sel(x)
}
\arguments{
\item{x}{An character vector containing the column numbers from the list}
}
\value{

}
\description{
Biophysical indicators selection function
}
\author{
Caio Faro
}
