% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indG_sel.R
\name{indG_sel}
\alias{indG_sel}
\title{Select indicators}
\usage{
indG_sel(x)
}
\arguments{
\item{x}{An character vector containing the column numbers from the list}
}
\value{

}
\description{
Governance indicators selection function
}
\author{
Caio Faro
}

