% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indS_sel.R
\name{indS_sel}
\alias{indS_sel}
\title{Select indicators}
\usage{
indS_sel(x)
}
\arguments{
\item{x}{An character vector containing the column numbers from the list}
}
\value{

}
\description{
Socioeconomical indicators selection function
}
\author{
Caio Faro
}

