% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings.R
\name{landings}
\alias{landings}
\title{Title}
\usage{
landings(data, location, type, species = NULL)
}
\arguments{
\item{data}{A data.frame with landings and income data}

\item{location}{The community of interest}

\item{type}{Specify if the function must calculate "price" for income or "kg" for landings}

\item{species}{An OPTIONAL species of interest}
}
\value{
A data.frame with the Landings or Income per year
}
\description{
Title
}

