% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{list_mpa}
\alias{list_mpa}
\title{List the locally stored MPAs}
\usage{
list_mpa(
  path = rappdirs::user_data_dir("wdpar"),
  ext = ".gpkg",
  full_names = FALSE
)
}
\arguments{
\item{path}{the path to storage}

\item{ext}{character, the file extension, by default we use ".gpkg"}

\item{full_names}{logical, if TRUE return full filenames}
}
\description{
List the locally stored MPAs
}
