% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{list_obis}
\alias{list_obis}
\title{List the locally stored OBIS datasets}
\usage{
list_obis(
  path = rappdirs::user_data_dir("robis"),
  ext = ".csv.gz",
  full_names = FALSE
)
}
\arguments{
\item{path}{the path to storage}

\item{ext}{character, the file extension, by default we use ".csv.gz"}

\item{full_names}{logical, if TRUE return full filenames}
}
\description{
List the locally stored OBIS datasets
}
