% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_area.R
\name{mpa_area}
\alias{mpa_area}
\title{area matrix generator}
\usage{
mpa_area(m, n, mpa.size, mpa.number, harvesting.rate = 0.1,
  harvesting.strategy = FALSE)
}
\arguments{
\item{m}{number of vertical cells (rows)}

\item{n}{number of horizontal cells (columns)}

\item{mpa.size}{proportion of cells to be MPAs (from 0 to 1)}

\item{mpa.number}{number of MPAs}

\item{harvesting.rate}{harvesting rate expressed as a proportion (0 to 1) of organisms harvested per cell. If harvesting.strategy = TRUE, this is the maximum value observed in the borders of the MPA.}

\item{harvesting.strategy}{A logical value that indicates if "Fising the line" occurs. harvesting.strategy=FALSE is the default that indicates equal harvesting rate outside the MPA. If TRUE, it creates a pattern in which harvesting rate depends on the distance to MPA.}
}
\value{
area A matriz of size m by n that simulates harvesting rates and MPAs
}
\description{
Generates the area matrix that will be simulated, based on size of the area, size of the MPA, number of MPAs,harvesting rate, and harvesting strategy.
}
\author{
Villasenor-Derbez, J.C.
}
