% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_check.R
\name{mpa_check}
\alias{mpa_check}
\title{mpa_sim input validator}
\usage{
mpa_check(area, nsteps, r, pop0, K, mrate)
}
\arguments{
\item{area}{an N by M matrix with harvesting rates for fishing grounds (0 < harvesting rate <= 1) and MPAs (harvesting rate = 0) for the simulated area.}

\item{nsteps}{Number of steps to run the simulation.}

\item{r}{Intrinsic population growth rate.}

\item{pop0}{Initial population size. If single, assumes equal population size across all cells in area. If a matrix, it must have the same dimensions as area.}

\item{K}{Carrying capacity}

\item{mrate}{Movement rate expressed as a proportion (relative to 1). Indicates the proportion of organisms that will move out of a cell in one timestep.}

\item{op}{A logical value that indicates if a filled.contour is to be created. WARNING: op = TRUE may make the simulation slow.}
}
\value{
check an object of type list containing elemetns result and message. Result is a logical that indicates if the inputs to mpa_2d() are correct. If they are not, message contains the error message to be displayed in the Console.
}
\description{
Validates the inputs passed to mpa_2D().
}
\author{
Villasenor-Derbez, J.C.
}

