% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_compare.R
\name{mpa_compare}
\alias{mpa_compare}
\title{Scneario concatenator}
\usage{
mpa_compare(results, names)
}
\description{
Creates a list containing results from multiple scenarios ran in mpa_sim
}
\seealso{
mpa_plot

mpa_sim
}
