% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es50.R
\name{mpa_es50}
\alias{mpa_es50}
\title{Compute ES50 for a sf table of MPAs}
\usage{
mpa_es50(mpa = read_mpa("Cuba"), verbose = interactive())
}
\arguments{
\item{mpa}{sf table of MPAs, possibly with a column of 'obs'.  If that column
is missing we add it by calling \code{fetch_obis} which could take some time.}

\item{verbose}{logical, output helpful messages}
}
\value{
the same mpa row with 'total_records', 'total_phyla', 'total_species', 
  'es50_overall' and possibly 'obs' if it doesn't already exist
}
\description{
Compute ES50 for a sf table of MPAs
}
