% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{mpa_knit_obs}
\alias{mpa_knit_obs}
\title{Given a set of MPAs and observations (like obis), add a list columns of 
obs to each row of MPAs that contain observations. Rows of mpa that have no
obs will have an empty tibble.}
\usage{
mpa_knit_obs(mpa = read_mpa("Cuba"), obs = fetch_obis("Cuba", form = "sf"))
}
\arguments{
\item{mpa}{sf table of MPAs}

\item{obs}{sf table of observations, like OBIS}
}
\value{
sf table with added 'obs' column
}
\description{
Given a set of MPAs and observations (like obis), add a list columns of 
obs to each row of MPAs that contain observations. Rows of mpa that have no
obs will have an empty tibble.
}
