% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_plot.R
\name{mpa_plot}
\alias{mpa_plot}
\title{Plots results of mpa_sim}
\usage{
mpa_plot(data, type = "io", in.col = "blue", in.type = "b",
  out.col = "red", out.type = "b", pop.col = "black", pop.type = "b",
  catch.col = "magenta", catch.type = "b", other = NULL, ...)
}
\description{
Plots results of mpa_sim
}
\seealso{
mpa_sim

mpa_compare
}
\author{
Villasenor-Derbez, J.C.
}
