% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_plot2.R
\name{mpa_plot2}
\alias{mpa_plot2}
\title{Plots results from biophysical indicator functions}
\usage{
mpa_plot2(data, type)
}
\arguments{
\item{data}{dataframe generated by any of the following functions: richness(), density(), trophic(), fish_biomass(), and fish_size()}

\item{type}{a quoted character that indicates which function was used to create the data frame. Use "s", "d", "t", "b", and "l", respectively.}
}
\description{
Plots results from the functions used to calculate biophysical indicators like: richness(), density(), ttrophic(), fish_biomass(), and fish_size().
}
\author{
Villasenor-Derbez, J.C.
}

