% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_plot3.R
\name{mpa_plot3}
\alias{mpa_plot3}
\title{Plots results from biophysical indicator functions by site}
\usage{
mpa_plot3(data, reserve = NULL, control = NULL, error.bars = F,
  y.lab = "Indicador")
}
\arguments{
\item{data}{dataframe generated by any of the following functions: richness(), density(), trophic(), fish_biomass(), and fish_size().}

\item{reserve}{A quoted string that indicates which site should be used as a reserve.}

\item{control}{A quoted string that indicates which site should be used as a control zone for the indicated reserve.}

\item{error.bars}{A logical value indicating if error bars (1 SD) should be plotted.}

\item{x.lab}{A string to be used as x-axis label}
}
\description{
Plots results from the functions used to calculate biophysical indicators like: richness(), density(), ttrophic(), fish_biomass(), and fish_size().
}
\author{
Villasenor-Derbez, J.C.
}

