% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_plot4.R
\name{mpa_plot4}
\alias{mpa_plot4}
\title{Plot a turfeffect object for biophysical indicators}
\usage{
mpa_plot4(model, y.lab = NULL)
}
\arguments{
\item{model}{An object of class lm, generated by the turfeffect() function or extracted from the tibble output of the bio_results() function}

\item{y.lab}{A character that specifies the y.label}
}
\value{
an object og class ggplot2
}
\description{
Plot a turfeffect object for biophysical indicators
}

