% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPAreport_doc.R
\name{MPAreport_doc}
\alias{MPAreport_doc}
\title{Create report for an MPA}
\usage{
MPAreport_doc(peces, invertebrados, comunidad, reserva, control)
}
\arguments{
\item{peces}{A data.frame}

\item{invertebrados}{A data.frame}

\item{comunidad}{A string}

\item{reserva}{A string}

\item{control}{A string}
}
\description{
Produces an html document with the analysis of 5 biophysical indicators performed for the given reserve/control site pair
}
