% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPAreport_html.R
\name{MPAreport_html}
\alias{MPAreport_html}
\title{Create report for an MPA}
\usage{
MPAreport_html(peces, invertebrados, pesca, gov, coop, comunidad, reserva,
  control)
}
\arguments{
\item{peces}{A data.frame}

\item{invertebrados}{A data.frame}

\item{comunidad}{A string}

\item{reserva}{A string}

\item{control}{A string}
}
\description{
Produces an html document with the analysis of 5 biophysical indicators performed for the given reserve/control site pair
}
