% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{obis_strict_match}
\alias{obis_strict_match}
\title{Ensure that a set of OBIS observations actually fall within a set of MPAs}
\usage{
obis_strict_match(
  obs = read_obis("Cuba"),
  mpa = read_mpa("Cuba"),
  form = c("tibble", "sf", "index")[1]
)
}
\arguments{
\item{obs}{tibble of OBIS observations}

\item{mpa}{sf MULTIPOLYGON object of MPAs form WDPA}

\item{form}{character one of
\itemize{
\item{tibble return a tibble}
\item{sf return an sf object (spatially referenced tibble)}
\item{index a list of indices that provide the match between obs and mpas}
}}
}
\value{
tibble of obs with possibly fewer rows
}
\description{
Ensure that a set of OBIS observations actually fall within a set of MPAs
}
\examples{
\dontrun{
  mpa = read_mpa("Belgium")
  obs = read_obis("Belgium")
  obs_strict <- obis_strict_match(obs, mpa)
  dim(obs)
  # [1] 511490    161
  dim(obs_strict)
  # [1] 480302    161
}
}
