% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{read_mpa}
\alias{read_mpa}
\title{Read a previously downloaded MPA dataset for a country, region, etc}
\usage{
read_mpa(name = "Cuba", path = rappdirs::user_data_dir("wdpar"), ext = ".gpkg")
}
\arguments{
\item{name}{character, the name of the country or 'global' that you have previously
downloaded}

\item{path}{character, the output path.  We default to \code{\link[rappdirs]{user_data_dir}}}

\item{ext}{character, the file extension, by default we use ".gpkg"}
}
\value{
a simple features table
}
\description{
Read a previously downloaded MPA dataset for a country, region, etc
}
\examples{
\dontrun{
  require(sf)
  x <- mpa_read("Cuba")
  plot(x['REP_AREA'])
}
}
