% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_table.R
\name{reg_table}
\alias{reg_table}
\title{Create a regression table}
\usage{
reg_table(model, title = "", dep.var.labels = "")
}
\arguments{
\item{model}{An object of class lm}

\item{title}{A title for the table}

\item{dep.var.labels}{Labels for dependent variables}
}
\description{
Create a regression table
}
\author{
Juan Carlos Villasenor
}
