% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_size.R
\name{reserve_size}
\alias{reserve_size}
\title{Title}
\usage{
reserve_size(data, spp, length, width)
}
\arguments{
\item{data}{}

\item{spp}{}

\item{length}{}

\item{width}{}
}
\value{
a list
}
\description{
Title
}
\examples{

}
