% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{Calculate species richness}
\usage{
richness(data, site)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Year, Zone (e.g. reserve or control site), Transect Number, GenusSpecies.}

\item{site}{A quoted string that indicates the site. Options are "Rosario", "IslaNatividad", "IslaMagdalena".}
}
\value{
S A dataframe with columns for Year, Zone (inside outside the reserve), Transect Number, and Species richness (S).
}
\description{
Calculates species density, by transect, for a specific location.
}

