% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveMAREA.R
\name{saveMAREA}
\alias{saveMAREA}
\title{Title}
\usage{
saveMAREA(data, comunidad, reserva)
}
\arguments{
\item{data}{data to store}

\item{comunidad}{community}

\item{reserva}{reserve}
}
\description{
Title
}
\examples{

}
