% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Shannon-Wiener Diversity Index}
\usage{
shannon(data, location)
}
\arguments{
\item{data}{A data.frame}

\item{location}{A string}
}
\value{
H A data.frame
}
\description{
Shannon-Wiener Diversity Index
}
