% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_results.R
\name{soc_results}
\alias{soc_results}
\title{Results}
\usage{
soc_results(values, data)
}
\arguments{
\item{values}{A list, created within the Shiny app, that contains the list of indicators selected and the community}

\item{data}{The data to be analyzed}
}
\value{
results, a tibble containing the name of the indicator as well as DiD estimates, p values, a string to be used as input for valueBox, the color of the valueBox, and the lm object fit to the indicator extracted from the data
}
\description{
Analyzes all the required biophysical indicators at once
}

