% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{sp_list}
\alias{sp_list}
\title{Title}
\usage{
sp_list(fish, invert, rc)
}
\arguments{
\item{fish}{a fish database}

\item{invert}{an invert database}

\item{rc}{the reserve - control pair}
}
\value{
sp_list
}
\description{
Title
}
\examples{

}
