% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic.R
\name{trophic}
\alias{trophic}
\title{Calculate mean trophic level}
\usage{
trophic(data, tl = NULL, site)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Ano, Zonificacion (e.g. "Zona de pesca" or "Zona de no pesca"), Transecto, GeneroEspecie, and Abundancia.}

\item{tl}{A dataframe with a column for each species (called GenusSpecies) and a column for the trophic level of each species (called TL).}

\item{location}{A quoted string that indicates the location.}
}
\value{
size A dataframe with columns for Ano, Zonificacion, Transect Number, Species, and mean trophic level.
}
\description{
Calulates mean trophic level, by year and transect, for each location.
}

