% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turfeffect.R
\name{turfeffect}
\alias{turfeffect}
\title{turfeffect}
\usage{
turfeffect(data, reserve = NULL, control = NULL, type = NULL)
}
\arguments{
\item{data}{A data.frame generated by any of the following functions: richness(), density(), trophic(), fish_biomass(), and fish_size().}

\item{reserve}{A string}

\item{control}{A string}
}
\value{
An object of class "lm"
}
\description{
turfeffect
}

