% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxString.R
\name{valueBoxString}
\alias{valueBoxString}
\title{valueBoxValues}
\usage{
valueBoxString(model, type = NULL)
}
\arguments{
\item{model}{An object of class lm, generated by the turfeffect function}

\item{type}{A string}
}
\value{
x, a list with the first element (x$x) containing the string of slope and p value, and a second element (x$color) containing the color that the valueBox should have
}
\description{
Calculate color and string needed in the ShinyApp valueBoxOutput based on a turfeffect model output
}
