% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{wdpa_cleaned_crs}
\alias{wdpa_cleaned_crs}
\title{Retrieve the projected CRS for WDPA cleaned (ESRI:54017) data}
\usage{
wdpa_cleaned_crs(form = c("proj4string", "epsg", "wkt")[3])
}
\arguments{
\item{form}{character one of 'proj4string', 'epsg', or 'wkt'}
}
\value{
character
}
\description{
Retrieve the projected CRS for WDPA cleaned (ESRI:54017) data
}
