% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{wdpa_fetch_country}
\alias{wdpa_fetch_country}
\title{Fetch WDPA data by country.  Be patient with this - it can take a while.}
\usage{
wdpa_fetch_country(
  country = "Cuba",
  path = rappdirs::user_data_dir("wdpar"),
  ext = ".gpkg",
  clean = TRUE,
  ...
)
}
\arguments{
\item{country}{character, either accepted country name or ISO-3 code. You can also
downlaod thge global dataset using \code{global}.}

\item{path}{character, the output path.  We default to that specified by
\code{\link[rappdirs]{user_data_dir}}}

\item{ext}{character, the file extension, by default we use ".gpkg"}

\item{clean}{logical, if TRUE run the resut through \code{\link[wdpar]{wdpa_clean}}}

\item{...}{other arguments for , see \code{\link[wdpar]{wdpa_fetch}}}
}
\value{
sf table of protected areas
}
\description{
You may get a warning about the "GDAL Error 1: ..." but it doesn't seem to be fatal and
the download is unaffected.
}
\examples{
\dontrun{
x <- wdpa_fetch_country(country = 'Cuba')
plot((x \%>\% dplyr::filter(MARINE == "1"))['REP_AREA'])
}
}
