% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{wdpa_read_country}
\alias{wdpa_read_country}
\title{Read a previously downloaded WDPA dataset for a country}
\usage{
wdpa_read_country(
  country = "Cuba",
  path = rappdirs::user_data_dir("wdpar"),
  ext = ".gpkg"
)
}
\arguments{
\item{country}{character, the name of the country or 'global' if you have previously
downloaded that datatset}

\item{path}{character, the output path.  We default to \code{\link[rappdirs]{user_data_dir}}}

\item{ext}{character, the file extension, by default we use ".gpkg"}
}
\value{
a simple features table
}
\description{
Read a previously downloaded WDPA dataset for a country
}
\examples{
\dontrun{
  require(sf)
  x <- wdpa_read_country("Cuba")
  plot(x['REP_AREA'])
}
}
