% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa.R
\name{which_geometry}
\alias{which_geometry}
\title{Determine the column that contains geometry in a table}
\usage{
which_geometry(x, form = c("index", "character")[1], pattern = "^geom")
}
\arguments{
\item{x}{table of data}

\item{form}{character either 'index' (default) or 'name'}

\item{pattern}{character the regular expression pattern to search with}
}
\description{
If more than one column matches the pattern then just the first is returned.
}
