% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{write_obis}
\alias{write_obis}
\title{Write an OBIS table to csv}
\usage{
write_obis(x, name, path = rappdirs::user_data_dir("robis"), overwrite = FALSE)
}
\arguments{
\item{x}{OBIS table of data}

\item{name}{character the name to save the dataset under (no spaces please!)}

\item{path}{character, the output path.  We default to that specified by
\code{\link[rappdirs]{user_data_dir}}}

\item{overwrite}{logical, if TRUE allow existing files to be overwritten}
}
\value{
the input data table
}
\description{
Write an OBIS table to csv
}
