% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2b.R
\name{a2b}
\alias{a2b}
\title{Convert data format}
\usage{
a2b(a)
}
\arguments{
\item{a}{A data.frame with the format a. Un data.frame en formato a}
}
\value{
A data.frame wih the format b. Un data.frame en formato b
}
\description{
Convert fish transect data from format a to format b; See details. Convierte datos de transectos de peces del formato a al formato b; Ver detalles
}
\examples{
#library(MPAtools)

#load an "a" sample dataset
data("a")

#convert a to b format
b <- a2b(a)
b

}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
