% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2a.R
\name{b2a}
\alias{b2a}
\title{Convert data format}
\usage{
b2a(b)
}
\arguments{
\item{b}{An object of class data.frame wih the format b (One column for sizes and one for abundance).}
}
\value{
a An object of class data.frame with the format a (one column for each size interval).
}
\description{

}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>
}

