% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c.R
\name{b2c}
\alias{b2c}
\title{Convert data format}
\usage{
b2c(b)
}
\arguments{
\item{a}{A data.frame with the format b. Un data.frame en formato b}
}
\value{
A data.frame wih the format c. Un data.frame en formato c
}
\description{
Convert fish transect data from format b to format c; See details. Convierte datos de transectos de peces del formato b al formato c; Ver detalles
}
\examples{
#library(MPAtools)

#load a "b" sample dataset
data("b")

#convert a to b format
c <- b2c(b)
c

}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
