% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2a.R
\name{c2a}
\alias{c2a}
\title{Convert data format}
\usage{
c2a(c)
}
\arguments{
\item{c}{An object of class data.frame wih the format c (One column for sizes and one for abundance).}
}
\value{
a An object of class data.frame with the format a (one column for each size interval).
}
\description{

}
\author{
Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>
}

