% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2b.R
\name{c2b}
\alias{c2b}
\title{Convert data format}
\usage{
c2b(c)
}
\arguments{
\item{a}{A data.frame with the format c. Un data.frame en formato c}
}
\value{
A data.frame wih the format b. Un data.frame en formato b
}
\description{
Convert fish transect data from format c to format b; See details. Convierte datos de transectos de peces del formato c al formato b; Ver detalles
}
\examples{
#library(MPAtools)

#load an "c" sample dataset
data("c")

#convert c to a format
b <- c2b(c)
b

}
\author{
Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>
}
