% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density}
\alias{density}
\title{Calculate density}
\usage{
density(data, site, species = NULL)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Year, Zone (e.g. reserve or control site), Transect Number, GenusSpecies.}

\item{site}{A quoted string that indicates the site. Options are "Rosario", "IslaNatividad", "IslaMagdalena".}

\item{species}{A quoted string that indicates a species for which density should be calulated. Common cases include "Paralabrax clathratus", "Semicossipjus pulcher", or "Anisotremus davidsonii".}
}
\value{
D A dataframe with columns for Year, Zone (inside outside the reserve), Transect Number, Species, and Density (org/m2).
}
\description{
Calculates species density, by year and transect, for a specific location.
}

