% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did}
\alias{did}
\title{Perform Difference-in-Difference analysis}
\usage{
did(data, year.imp, year.int = NULL)
}
\arguments{
\item{data}{A dataframe that has been created with any of the following functions: richness(), density(), fish_biomass(), fish_size(), trophic(). All these functions are part of this package.}

\item{year.imp}{An integer that indicates the year when the reserve was first implemented.}

\item{year.int}{An integer that indicates the year of interest to perform the DiD analysis}
}
\value{
Returns an objet of class "lm", where the value of the interactive term between Zonificacion and Ano (Zonificacion and Ano) is the difference-in-difference estimate.
}
\description{
Performs a difference-in-difference analysis on a series of data and compares the effect of the reserve. It is based on a linear model, where the variable of interest is regressed on dummy variables composed by Ano and Zonificacion, and an interaction term between both. The model is OutcomeVariable = b0 + b1*Ano + b2*Zonificacion + b3(Ano*Zonificacion). Here, Ano indicates pre and post implementation of the reserve. Zonificacion serves as a a way to compare a trend and allows to compare between the reserve and control sites. The DD value is indicated by the coefficient of the interaction term b3.
}

