% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es50.R
\name{es50_base}
\alias{es50_base}
\title{Record number of records and attempt an overall es50 calculation}
\usage{
es50_base(mpa, key, verbose = interactive())
}
\arguments{
\item{mpa}{a single row of an mpa records table}

\item{key}{the wdpaid of the mpa}

\item{verbose}{logical, if TRUE output helpful messages}
}
\value{
the same mpa row with total_records, total_phyla, total_species, es50_overall
 or NULL if an error occurs
}
\description{
Record number of records and attempt an overall es50 calculation
}
\seealso{
\code{\link{mpa_es50}}
}
