% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es50.R
\name{es50_timeblock}
\alias{es50_timeblock}
\title{Calculate es50 metric for an mpa for specific time points}
\usage{
es50_timeblock(mpa, age, step)
}
\arguments{
\item{mpa}{a row from wdpa table containing one mpa with metadata and geometry column}

\item{age}{numeric, age to compute es50 for}

\item{records}{an object containing obis records}

\item{step_size}{numeric of years in each step of analysis}
}
\value{
single row table containing same columns as mpa but with added timeblock columns
}
\description{
Calculate es50 metric for an mpa for specific time points
}
