% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{fetch_obis}
\alias{fetch_obis}
\title{Given a simple features tibble (POLYGON or MUTLIPOLYGON), retrieve OBIS observation records.}
\usage{
fetch_obis(x = read_mpa("Cuba"), policy = c("convex hull", "strict")[1])
}
\arguments{
\item{x}{MPA sf object such as for a country}

\item{policy}{character either "convex hull" or "strict"
We fetch the convex hull to simplify the fetch, but that may allow some observations that are outside MPAs.
If "convex hull" then we allow these to be returned.  If "strict" then we filter for those observations that are in
the MPAs (in mean within or on the boundary).}
}
\value{
tibble of OBIS observations, possibly empty
}
\description{
Given a simple features tibble (POLYGON or MUTLIPOLYGON), retrieve OBIS observation records.
}
