% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{fetch_obis_country}
\alias{fetch_obis_country}
\title{Fetch OBIS data by country}
\usage{
fetch_obis_country(name = "Cuba", overwrite = TRUE, ...)
}
\arguments{
\item{name}{character, the name to save the MPA under (no spaces please!)}

\item{...}{other arguments for obis_fetch_mpa}

\item{mpa}{POLYGON or MULTIPOLYGON as sf}

\item{save_file}{logical, if TRUE save into the country's data folder}
}
\value{
tibble of OBIS observations, possibly empty
}
\description{
No error checking is done to ensure the country folderand WDPA file
}
