% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_biomass.R
\name{fish_biomass}
\alias{fish_biomass}
\title{Calculate bbiomass}
\usage{
fish_biomass(data, site, species = NULL)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Year, Zone (e.g. reserve or control site), Transect Number, GenusSpecies.}

\item{site}{A quoted string that indicates the site. Options are "Rosario", "IslaNatividad", "IslaMagdalena".}

\item{species}{A quoted string that indicates a species for which density should be calulated. Common cases include "Paralabrax clathratus", "Semicossipjus pulcher", or "Anisotremus davidsonii".}
}
\value{
size A dataframe containing year, zone (insido or outside the reserve), transect number, species, and the statistic required.
}
\description{
Calulates biomass for a selected species or the entire dataset, by year and transect, for each location.
}

