% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_size.R
\name{fish_size}
\alias{fish_size}
\title{Calculate a descriptor for fish size}
\usage{
fish_size(data, location, stat = "mean", species = "")
}
\arguments{
\item{data}{A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie, Abundancia.}

\item{location}{A quoted string that indicates the location.}

\item{stat}{A quoted string that indicates which statistic should be calculated. Options include "max", "min", "mean", and "median". User can also use "all" to calculate all of the above.}

\item{species}{A quoted string that indicates a species for which the size statistic should be calulated.}
}
\value{
size A dataframe with columns for Ano, Zona, Sitio, Transecto, GeneroEspecie, and the statistic required.
}
\description{
Calulates mean, median, maximum, or minimum size (cm) for a list of species (or selected species, as indicated by parameter species) by year, site, and transect, for a speciic location.
}

