% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_plot4.R
\name{mpa_plot4}
\alias{mpa_plot4}
\title{Title}
\usage{
mpa_plot4(model, y.lab = NULL)
}
\arguments{
\item{model}{An object of class lm, generated by the turfeffect() function or extracted from the tibble output of the bio_results() function}
}
\description{
Title
}

