% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{obis_as_sf}
\alias{obis_as_sf}
\title{Cast an OBIS data frame as a simple feature (POINT)}
\usage{
obis_as_sf(x, crs = c(NA, wdpa_cleaned_crs("wkt"))[2], ...)
}
\arguments{
\item{x}{tibble of OBIS data}

\item{crs}{WKT, by default wdpa_cleaned_crs("wkt") or NA to skip}

\item{...}{other arguments for \code{\link[sf]{st_as_sf}}}
}
\value{
sf object (POINT)
}
\description{
Cast an OBIS data frame as a simple feature (POINT)
}
