% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{Calculate species richness}
\usage{
richness(data, location)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie.}

\item{location}{A quoted string that indicates the location.}
}
\value{
S A dataframe with columns for Ano, Zona, Sitio, Transecto, and Species richness (S; number of species).
}
\description{
Calculates species richness, by year, site, and transect, for a specific location.
}

