% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_plot.R
\name{soc_plot}
\alias{soc_plot}
\title{Plot a turfeffect object for socioeconomic indicators}
\usage{
soc_plot(data, y.lab = "Value")
}
\arguments{
\item{y.lab}{A character that specifies the y label}

\item{model}{An object of class lm, generated by the turfeffect() function or extracted from the tibble output of the bio_results() function}
}
\value{
an object of class ggplot2
}
\description{
Plot a turfeffect object for socioeconomic indicators
}
