% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic.R
\name{trophic}
\alias{trophic}
\title{Calculate mean trophic level}
\usage{
trophic(data, site)
}
\arguments{
\item{data}{A dataframe that contains at least columns of Year, Zone (e.g. reserve or control site), Transect Number, GenusSpecies.}

\item{site}{A quoted string that indicates the site. Options are "Rosario", "IslaNatividad", "IslaMagdalena".}
}
\value{
size A dataframe containing year, zone (insido or outside the reserve), transect number, species, and the statistic required.
}
\description{
Calulates mean trophic level, by year and transect, for each location.
}

